move files to respective folders and fix imports
