+++ src/pages/App/App.css
@@ -1,4 +1,4 @@
-.App {
+/* .App {
   text-align: center;
 }
 
@@ -35,4 +35,4 @@
   to {
     transform: rotate(360deg);
   }
-}
+} */
