+++ src/pages/App/App.js
@@ -1,0 +1,11 @@
+import './App.css';
+
+function App() {
+  return (
+    <>
+      <div>Genshin Builds</div>
+    </>
+  );
+}
+
+export default App;
